Fix extra decref
